requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain_community
langchain_experimental
langchain_huggingface
langchain_groq
groq
langchain_chroma
pypdf
python-dotenv
pdfminer.six
